requirements.txt
google-adk>=1.4.1
google-generativeai>=0.8.5
requests>=2.32.4

